fixed new SoundFont Unit Tests

+++ app/src/main/java/com/qfs/apres/soundfont/Generator.kt
@@ -21,8 +21,8 @@ class Generator(
         }
     }
     fun asTimecent(): Double {
-        val p = this.asIntSigned() / 1200
-        return (2.0).pow(p.toDouble())
+        val p = this.asIntSigned().toDouble() / 1200.toDouble()
+        return (2.0).pow(p)
     }
     fun asPair(): Pair<Int, Int> {
         return Pair(this.shAmount, this.wAmount)
